+++ src/finaletoolkit/utils/typing.py
@@ -1,5 +1,4 @@
 """
-# _typing
 Some useful type aliases
 """
 from __future__ import annotations
@@ -10,4 +9,5 @@ from pysam import AlignmentFile, TabixFile
 
 # files accepted by frag_generator
 FragFile = Union[str, PathLike, AlignmentFile, TabixFile]
+ChromSizes = Union[str, PathLike]
 Intervals = Union[str, PathLike]